More refactor
